addProjectRole

+++ app/Http/Controllers/CompanyProjectController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Company;
 use App\Models\Project;
+use App\Models\ProjectRole;
+use Illuminate\Contracts\Support\ValidatedData;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Validator;
 
@@ -26,8 +30,8 @@ class CompanyProjectController extends Controller
         return ResponseFormatter::error(null, 'Project not found', 404, "{$id} not found");
       }
     }
-
-    $project = Project::with(['company'])->latest();
+    // return Auth::user();
+    $project = Project::where('company_id', Auth::user()->company_id)->latest();
 
     return ResponseFormatter::success($project->paginate($limit), "Data Successfully Retrieved");
   }
@@ -53,7 +57,7 @@ class CompanyProjectController extends Controller
       $validatedData['images'] =  "\images\project\\" . "default.png";
     }
 
-    $validatedData["company_id"] = 0;
+    $validatedData["company_id"] = Auth::user()->company_id;
 
     $project = Project::create($validatedData);
 
@@ -77,6 +81,9 @@ class CompanyProjectController extends Controller
     $project = Project::find($id);
 
     if ($project) {
+      if (Auth::user()->company_id != $project->company_id) {
+        return ResponseFormatter::error(null, 'You are not in this company', 401, "You are not in this company");
+      }
       if ($request->hasFile('image')) {
         File::exists(public_path() . $project->images) && $project->images != '/images/project/default.png' ? File::delete(public_path() . $project->images) : '';
         $file = $request->file('image');
@@ -93,6 +100,35 @@ class CompanyProjectController extends Controller
     } else {
       return ResponseFormatter::error(null, "Project not found", 404, "Project {$id} not found");
     }
+  }
+
+  public function addProjectRole(Request $request, $id)
+  {
+    // return $request['extra-question'];
+    $project = Project::find($id);
+
+    if (Auth::user()->company_id != $project->company_id) {
+      return ResponseFormatter::error(null, 'You are not in this company', 401, "You are not in this company");
+    }
+
+    $validator = Validator::make($request->all(), [
+      'name' => 'required|string|max:255',
+      'quota' => 'required|integer',
+      'description' => 'required',
+    ]);
+
+    if ($validator->fails()) {
+      return ResponseFormatter::error(null, 'Validation Error', 400, $validator->errors());
+    }
+
+    $data = $request->all();
+
+    $data['project_id'] = $id;
+
+    // return $data;
+
+    $projectRole = ProjectRole::create($data);
+    return ResponseFormatter::success(ProjectRole::find($projectRole->id), "Project Role has been created");
   }
 
   public function getAllApplicants(Request $request)
